(PC-17806) feat(auth): move useDepositAmountsByAge to share/user folder
